[#31420] File re-organization
