+++ resources/js/components/Concept/Default.vue
@@ -144,399 +144,10 @@
 </template>
 
 <script>
-import { BButton, BInputGroup, BInputGroupAppend } from 'bootstrap-vue';
-import state from '../../states/concept';
-import EditableTerm from '../Term/Editable.vue';
-import EditableSource from '../Source/Editable.vue';
-import EditableCategory from '../Category/Editable.vue';
-import { categories } from '../../config/categories';
-import conceptApi from '../../api/ConceptService';
-import termApi from '../../api/TermService';
-import src from 'vue-multiselect/src';
+import Component from './Default.js';
 
-export default {
-  components: { BInputGroup, BInputGroupAppend, EditableTerm, EditableSource, EditableCategory },
-  props: {
-    conceptProps: {
-      type: Object,
-    },
-    termProps: {
-      type: Array,
-    },
-    categoriesProps: {
-      type: Array,
-    },
-    sourcesProps: {
-      type: Array,
-    },
-    canEditVocabulary: false,
-  },
-  data() {
-    return {
-      deprecated: this.conceptProps.deprecated,
-      // Do we need to make a copy of termProps with slice() or [...this.termProps]?
-      terms: this.termProps.map(
-        // populating terms with our custom temporary variables
-        (term,index) => {
-          term.inEdit = false;
-          term.index = index;
-          return term;
-        },
-      ),
-      sources: this.sourcesProps.map(
-        // populating terms with our custom temporary variables
-        (source) => {
-          source.inEdit = false;
-          return source;
-        },
-      ),
-      cats: this.categoriesProps,
-      conceptId: this.conceptProps.id,
-      termSearch: [],
-      allTermsSearch: false,
-      selected_concept: '',
-      relationType: '',
-      state: state,
-      categories,
-      isVocabularyEditor: this.canEditVocabulary !== 'false',
-      baseURL: process.env.MIX_APP_URL,
-    };
-  },
-  computed: {
-    src() {
-      return src;
-    },
-    alternateTerms() {
-      return this.terms
-        .filter((term) => !term.preferred)
-        .sort();
-    },
-    preferredTerm() {
-      return this.terms.find((term) => term.preferred);
-    },
-    hasEmptySource() {
-      return !!(
-        this.sources.length && !this.sources[this.sources.length - 1].id
-      );
-    },
-    selectedCategories() {
-      return this.cats.map((cat) => cat.id);
-    },
-    hasEmptyCategory() {
-      return !!(
-        this.cats.length &&
-        !this.cats[
-          this.cats.length - 1
-        ].id
-      );
-    },
-    hasEmptyTerm() {
-      return !!(
-        this.terms.length &&
-        !this.terms[
-          this.terms.length - 1
-        ].text
-      );
-    },
-  },
-  methods: {
-    fetchConcept: function () {
-      fetch(`${this.baseURL}/api/concepts/` + this.terms[0].concept_id)
-        .then((data) => data.json())
-        .then((data) => {
-          console.log('here you are: ', data);
-          this.terms = data.terms.map(
-            // populating terms with our custom temporary variables
-            (term) => {
-              term.inEdit = false;
-              return term;
-            },
-          );
-        });
-    },
-    addTerm: function () {
-      if (this.hasEmptyTerm) {
-        return;
-      }
-
-      const conceptID = this.terms[0].concept_id;
-      const newTerm = {
-        concept_id: conceptID,
-        id: null,
-        preferred: false,
-        text: null,
-        inEdit: true,
-        index: this.terms.length,
-      };
-      this.terms.push(newTerm);
-      console.log('newTerms',this.terms);
-    },
-    async saveTerm(term, termIndex) {
-      const finalize = (term, termIndex) => {
-        term.inEdit = false;
-        if(termIndex > -1) {
-          this.terms.splice(termIndex, 1, term);
-        }
-        this.cleanDirty(term);
-        this.flashSuccessAlert();
-      };
-
-      if (!term.id) {
-        term = await this.createTerm(term);
-        finalize(term, termIndex);
-        return;
-      }
-
-      const [error,response] = await termApi.updateTerm(term.id, term);
-      if(!error) {
-        finalize(term, termIndex);
-      }
-    },
-    async createTerm (term) {
-      const [error,response] = await termApi.createTerm(term);
-      if(!error){
-        return response;
-      }
-      return term;
-    },
-    makeTermPreferred: function (term, termIndex) {
-      if (
-        !confirm(
-          `Are you sure you want to make '${term.text}' the preferred term for this concept?`,
-        )
-      ) {
-        return;
-      }
-
-      const currentPreferred = this.preferredTerm;
-
-      if (currentPreferred && currentPreferred.id !== term.id) {
-        currentPreferred.preferred = false;
-        this.saveTerm(currentPreferred, currentPreferred.index);
-      }
-
-      term.preferred = true;
-      this.saveTerm(term, termIndex);
-    },
-    async deleteTerm (termId, termText) {
-      if (!confirm('Are you sure you want to delete this term?')) {
-        return;
-      }
-
-      let term = {
-        id: termId,
-        text: termText,
-        concept_id: this.conceptId,
-        preferred: false,
-      };
-
-      let index = this.terms.findIndex((t) => t.id === termId);
-      if (index === -1) {
-        index = this.terms.findIndex((t) => t.text === termText);
-      }
-      if(index > -1) {
-        term = this.terms[index];
-      }
-
-      const finalize = (term, index) => {
-        this.terms.splice(index, 1);
-        this.cleanDirty(term);
-      };
-
-      if (!termId) {
-        finalize(term, index);
-        return;
-      }
-
-      const [error,response] = await termApi.deleteTerm(termId);
-      if (!error) {
-        finalize(term, index);
-      }
-    },
-    getCategories(selectedId) {
-      return this.categories.filter((cat) => {
-        return (
-          this.cats.filter((currentCat) => {
-            return currentCat.id === parseInt(cat.value);
-          }).length === 0 || selectedId === parseInt(cat.value)
-        );
-      });
-    },
-    async updateCategories() {
-      this.conceptProps.concept_categories = this.cats;
-      const [error,response] = await conceptApi.updateConcept(this.conceptId, {
-        conceptCategories: this.cats,
-      });
-      if(!error) {
-        this.flashSuccessAlert();
-      }
-    },
-    addCategory: function () {
-      if (this.hasEmptyCategory) {
-        return;
-      }
-
-      const newCategory = {
-        pivot: {
-          concept_id: this.conceptId,
-        },
-        type: 'concept_category',
-      };
-      this.cats.push(newCategory);
-    },
-    saveCategory: function (categoryId, index) {
-      console.log(`Saving Category with id ${categoryId}`);
-      const categoryText = this.categories.find((cat) => parseInt(cat.value) === categoryId).text;
-      const category = { ...this.cats[index], id: categoryId, value: categoryText };
-
-      this.$set(this.cats, index, category);
-
-      this.cleanDirty({ id: categoryId });
-
-      this.updateCategories();
-    },
-    deleteCategory: function(categoryId, index) {
-      console.log(`Deleting Category with id ${categoryId}`);
-
-      this.cats.splice(index, 1);
-
-      this.updateCategories();
-    },
-    addSource: function () {
-      this.sources.push({ concept_id: this.conceptId, inEdit: true });
-    },
-    updateSources: function (source, index) {
-      this.sources.splice(index, 1, source);
-    },
-    saveSource: function (source, sourceId, index) {
-      console.log(
-        `Saving ${sourceId} ConceptId: ${this.conceptId}, Index: ${index}`,
-      );
-      const vm = this;
-      if (sourceId) {
-        axios
-          .patch(`${this.baseURL}/api/concept_sources/${sourceId}`, source)
-          .then(function (response) {
-            vm.cleanDirty(response.data);
-            vm.updateSources(response.data, index);
-            vm.flashSuccessAlert();
-          })
-          .catch(function (error) {
-            console.log(error);
-          });
-      } else {
-        axios
-          .post(`${this.baseURL}/api/concept_sources`, source)
-          .then(function (response) {
-            vm.cleanDirty(source);
-            vm.updateSources(response.data, index);
-            vm.flashSuccessAlert();
-          })
-          .catch(function (error) {
-            console.log(error);
-          });
-      }
-    },
-    deleteSource: function (sourceId, index) {
-      console.log(`Deleting Source with id ${sourceId}`);
-
-      if (!sourceId) {
-        this.cleanDirty(this.sources[index]);
-        this.sources.splice(index, 1);
-        return;
-      }
-
-      const vm = this;
-
-      axios
-        .delete(`${this.baseURL}/api/concept_sources/${sourceId}`)
-        .then(function (response) {
-          console.log('Deleted!', response);
-          vm.cleanDirty(vm.sources[index]);
-          vm.sources.splice(index, 1);
-        })
-        .catch(function (error) {
-          console.log(error);
-        });
-    },
-    toggleEditMode: function () {
-      if (this.getEditMode() && this.isDirty()) {
-        if (
-          !confirm(
-            'You have unsaved changes. Are you sure you want to exit Edit Mode?',
-          )
-        ) {
-          return;
-        }
-        this.resetDirty();
-      }
-
-      this.state.editMode = !this.state.editMode;
-    },
-    getEditMode: function () {
-      return this.state.editMode;
-    },
-    flashSuccessAlert: function () {
-      const alert = document.querySelector('.alert-success');
-      alert.classList.remove('hidden');
-      setTimeout(function () {
-        alert.classList.add('hidden');
-      }, 3000);
-    },
-    isDirty: function () {
-      return this.state.isDirty.length > 0;
-    },
-    flagDirty: function (obj) {
-      if (obj.dirty) {
-        this.markDirty(obj);
-      } else {
-        this.cleanDirty(obj);
-      }
-    },
-    markDirty: function (obj) {
-      if (obj.id) {
-        for (let i = 0; i < this.state.isDirty.length; i++) {
-          if (this.state.isDirty[i].id === obj.id) {
-            return;
-          }
-        }
-      } else {
-        let prev = obj.previous;
-        delete obj.previous;
-        delete obj.dirty;
-
-        for (let i = 0; i < this.state.isDirty.length; i++) {
-          if (JSON.stringify(this.state.isDirty[i]) === JSON.stringify(prev)) {
-            this.state.isDirty[i] = obj;
-            return;
-          }
-        }
-      }
-      this.state.isDirty[this.state.isDirty.length] = obj;
-    },
-    cleanDirty: function (obj) {
-      delete obj.previous;
-      delete obj.dirty;
-
-      for (let i = 0; i < this.state.isDirty.length; i++) {
-        if (this.state.isDirty[i].id && this.state.isDirty[i].id === obj.id) {
-          this.state.isDirty.splice(i, 1);
-          return;
-        } else if (
-          JSON.stringify(this.state.isDirty[i]) === JSON.stringify(obj)
-        ) {
-          this.state.isDirty.splice(i, 1);
-          return;
-        }
-      }
-    },
-    resetDirty: function () {
-      this.state.isDirty = [];
-    },
-  },
-};
+export default Component;
 </script>
-
 <style scoped>
 b-input-group {
   min-width: 500px;
